Move document helpers to hadron document from compass-crud
